Remove frontend config
We don't use React or TypeScript.

Webpack is not necessary to bundle the one JS file in this app, as even
Wagtail's Webpack config only copies the file to the build directory.

Linting configuration is kept as we can still run it via pre-commit. CSS
linting is also kept, as even though we don't have any CSS files right
now, it might be handy in case we need CSS overrides to maintain
compatibility with latest Wagtail versions.