Add wraps to mocked function in tests to prevent RuntimeWarning
Without this, the following:
RuntimeWarning: Missing dependency 'author' for prepopulated_field 'title'
is shown when running the tests, because the mocked function no longer returns the
form fields correctly. We add wraps so that the logic in the original function
still runs, while retaining the ability to assert the call parameters.